requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Image processing
Pillow>=10.3.0
pillow-heif>=0.16.0
rawpy>=0.19.0

# EXIF
piexif>=1.1.3

# Numerics (embedding similarity)
numpy>=1.26.0

# Async HTTP (Nominatim reverse geocoding)
httpx>=0.27.0

# Task scheduling (backup periodico DB, re-scan)
apscheduler>=3.10.4

# Multipart form data e file upload
python-multipart>=0.0.9

# File I/O asincrono
aiofiles>=23.2.1

# AI engines
google-generativeai>=0.5.0
ollama>=0.2.0

# Auth (Fase 4)
authlib>=1.3.0
itsdangerous>=2.2.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
